Use dialog-based EasyOCR script
